AREG-68 fix the user service keycloak service
