+++ applications/portal/backend/api/services/user_service.py
@@ -1,5 +1,6 @@
 import jwt
 from cloudharness.middleware import get_authentication_token
+from api.services.keycloak_service import KeycloakService
 
 class UnrecognizedUser(Exception):
     pass
@@ -11,3 +12,13 @@ def get_current_user_id() -> str:
         return jwt.decode(token, options={"verify_signature": False}, algorithms='RS256')['sub']
     except Exception as e:
         raise UnrecognizedUser()
+    
+def check_if_user_is_admin():
+    """
+    Function that is used to validate the token and check if the user is an admin
+    """
+    auth = KeycloakService()
+    if auth.current_user_has_realm_role("administrator"):
+        return {}
+    raise Exception("User is not an admin")
+
